requirements.txt
Flask==2.3.3
beautifulsoup4==4.12.3
requests==2.31.0
textblob==0.17.1
matplotlib==3.8.4
seaborn==0.13.2
wordcloud==1.9.3
IMDbPY
plotly==5.23.0
better-profanity==0.7.0
spacy==3.7.4
python-dotenv==1.0.1
google-api-python-client==2.129.0
nltk==3.9.1